fix placeholder integration test
