+++ tests/integration/test_guidellm.py
@@ -1,5 +1,8 @@
-import guidellm
+import pytest
+
+from guidellm.config import settings
 
 
+@pytest.mark.smoke()
 def test_import():
-    assert guidellm
+    assert settings
